+++ tsdef/VideoTrack.d.ts
@@ -1,4 +1,4 @@
-import { Track } from './Track';
+import { MediaTrack } from './MediaTrack';
 import { VideoProcessor } from './VideoProcessor';
 
 export namespace VideoTrack {
@@ -8,22 +8,12 @@ export namespace VideoTrack {
   }
 }
 
-export class VideoTrack extends Track {
-  isStarted: boolean;
-  isEnabled: boolean;
+export class VideoTrack extends MediaTrack {
   dimensions: VideoTrack.Dimensions;
   kind: 'video';
   processor: VideoProcessor | null;
-  mediaStreamTrack: MediaStreamTrack;
   processedTrack: MediaStreamTrack | null;
 
   addProcessor(processor: VideoProcessor): this;
   removeProcessor(processor: VideoProcessor): this;
-  attach(element?: HTMLMediaElement | string): HTMLVideoElement;
-  detach(element?: HTMLMediaElement | string): HTMLVideoElement[];
-
-  on(event: 'disabled', listener: (track: this) => void): this;
-  on(event: 'enabled', listener: (track: this) => void): this;
-  on(event: 'started', listener: (track: this) => void): this;
-  on(event: string, listener: (...args: any[]) => void): this;
 }