+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The Twilio Programmable Video SDKs use [Semantic Versioning](http://www.semver.o
 - Fixed a bug that manifests on Chrome versions 112+ where `Room.getStats()` rejects the returned Promise due to a
   TypeError caused by trying to read from the now-removed `RTCMediaStreamTrackStats`. Instead, the SDK now reads from
   the `RTCMediaSourceStats`. (VIDEO-12411) 
+- Fixed an error in the type definition for the `attach()` method of `AudioTrack` and `VideoTrack`. (VIDEO-12242)
 - Fixed an error in the type definition for `createLocalAudioTrack()`. (VIDEO-12383)
 
 2.26.0 (December 14, 2022)